refactor: httpClient 클래스 사용 및 카테고리 타입 폴더 이동
