--- src/app/_home/types.ts
@@ -1,6 +1,0 @@
-export type CategoryType = {
-  id: number;
-  name: string;
-  createdAt: Date;
-  updatedAt: Date;
-};
